refactor: 图片/截图/浏览器存储 IO 迁移到 infra 并修复名片缓存发送
